requirements.txt
streamlit==1.18.0
langchain==0.16.0
langchain-core==0.16.0
langchain-community==0.0.2
langchain-groq==0.0.1
langchain-text-splitters==0.0.1
langchain-embeddings==0.1.0
faiss-cpu==1.7.2
pdfplumber==0.8.0
python-dotenv==1.0.0
